# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion de biens immobilier",
        "description": "Documentation de l'API immoApp",
        "version": "1.0.0"
    },
    "paths": {
        "/api/avis": {
            "get": {
                "tags": [
                    "Avis"
                ],
                "summary": "Liste tous les avis",
                "operationId": "074b394984977e6dba6974626b17ff22",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Avis"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Avis"
                ],
                "summary": "Créer un avis",
                "operationId": "c6eeaa4be5353dc347263747d29b964d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Avis"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Avis créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Avis"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/avis/{id}": {
            "get": {
                "tags": [
                    "Avis"
                ],
                "summary": "Afficher un avis",
                "operationId": "4167dc39a58a676e2b99040ac98ae715",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Avis trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Avis"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Avis"
                ],
                "summary": "Mettre à jour un avis",
                "operationId": "d1cd0f2a587d195a4e36eb1ad258f878",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Avis"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Avis mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Avis"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Avis"
                ],
                "summary": "Supprimer un avis",
                "operationId": "c158d9539417f8b9af2ad21eec4d7f98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Avis supprimé"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/biens": {
            "get": {
                "tags": [
                    "Biens"
                ],
                "summary": "Liste de tous les biens",
                "operationId": "c963b1c217a3a93e6e25c8517f59f6d3",
                "responses": {
                    "200": {
                        "description": "Liste retournée avec succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Biens"
                ],
                "summary": "Créer un nouveau bien",
                "operationId": "40198ae6ffff58efac4dbfdd6b24c362",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Biens"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bien créé avec succès"
                    }
                }
            }
        },
        "/api/biens/{id}": {
            "get": {
                "tags": [
                    "Biens"
                ],
                "summary": "Afficher un bien par son ID",
                "operationId": "3855f805820761cbb78cd890a445e067",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du bien",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bien trouvé avec succès"
                    },
                    "404": {
                        "description": "Bien non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Biens"
                ],
                "summary": "Mettre à jour un bien existant",
                "operationId": "009dc45601d8a78ca5e071adb9e25a49",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du bien",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Biens"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bien mis à jour avec succès"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Biens"
                ],
                "summary": "Supprimer un bien",
                "operationId": "637f477673ef6ed2561e4f69a05347d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du bien",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Bien supprimé avec succès"
                    }
                }
            }
        },
        "/api/categorie_biens": {
            "get": {
                "tags": [
                    "CategorieBiens"
                ],
                "summary": "Lister catégories",
                "operationId": "f2a4fa0d328368d571f467308348926c",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "CategorieBiens"
                ],
                "summary": "Créer catégorie",
                "operationId": "10630faa2dba82bb3f1a237fd116e47c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé"
                    }
                }
            }
        },
        "/api/categorie_biens/{id}": {
            "get": {
                "tags": [
                    "CategorieBiens"
                ],
                "summary": "Afficher catégorie",
                "operationId": "909b71a7687b3164407ed2f3335283c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails"
                    }
                }
            },
            "put": {
                "tags": [
                    "CategorieBiens"
                ],
                "summary": "Modifier catégorie",
                "operationId": "5b6899d03eb985cdce56f43e932da7ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Modifié"
                    }
                }
            },
            "delete": {
                "tags": [
                    "CategorieBiens"
                ],
                "summary": "Supprimer catégorie",
                "operationId": "94432bc7df6c0f3d0fd79e76a59b7669",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Avis": {
                "required": [
                    "utilisateur_id",
                    "commentaire",
                    "note"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "utilisateur_id": {
                        "type": "integer"
                    },
                    "commentaire": {
                        "type": "string"
                    },
                    "note": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Biens": {
                "title": "Biens",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "Jean"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Dupont"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jean.dupont@example.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}